refactor: AuthController, UserController class
UserController에 있는 로그인 / 로그아웃 인증 로직 AuthController로 분리
