+++ src/main/java/techcourse/myblog/exception/UserForbiddenException.java
@@ -1,0 +1,7 @@
+package techcourse.myblog.exception;
+
+public class UserForbiddenException extends RuntimeException {
+    public UserForbiddenException(final String message) {
+        super(message);
+    }
+}
